requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pymilvus>=2.3.0
transformers>=4.35.0
torch>=2.0.0
Pillow>=10.0.0
python-multipart>=0.0.6
